# openapi.yaml
openapi: 3.0.3
info:
  title: OLIM Backend API
  version: 0.1.0
  description: >
    Spécification OpenAPI (Swagger) générée à partir des routes/contrôleurs présents dans `src/`.
servers:
  - url: https://olim-backend.onrender.com
    description: Production (Render)
  - url: http://localhost:3000
    description: Local (par défaut)
...
paths:
...
  /v1/requests:
    post:
      tags: [Requests]
      summary: Créer une demande (backend-first)
      description: >
        Crée une demande dans Firestore (`Clients/{uid}/Requests/{requestId}`) puis tente un upsert dans Supabase
        (`public.requests`) avec une clé idempotente. Les dates ISO dans `request` sont acceptées et converties
        côté serveur si possible.
      security:
        - bearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          description: >
            Clé d’idempotence (recommandée). Alternative acceptée: `idempotencyKey` dans le body.
          schema:
            type: string
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [source, idempotencyKey, request]
              properties:
                source:
                  type: string
                  enum: [APP, VOICE, SHARE, SYSTEM]
                idempotencyKey:
                  type: string
                  maxLength: 128
                request:
                  type: object
                  description: Payload legacy Firestore (sans FieldValue, dates en ISO).
                  additionalProperties: true
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [requestId]
                properties:
                  requestId:
                    type: string
                    description: ID du document Firestore créé (docId)
        '400':
          description: Paramètres invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Requête déjà en cours (idempotency lock)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/Error'
...
  /api/requests:
    get:
      tags: [Requests]
      summary: Lister les demandes du client
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
          description: Filtre sur le champ legacy `Status`
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: Liste
          content:
            application/json:
              schema:
                type: object
                required: [requests]
                properties:
                  requests:
                    type: array
                    items:
                      $ref: '#/components/schemas/ClientRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/Error'
    post:
      tags: [Requests]
      summary: Créer une demande
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRequestInput'
      responses:
        '201':
          description: Créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: Non authentifié / token invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missing:
              value: { error: "Missing or invalid authorization header" }
            invalid:
              value: { error: "Invalid or expired token" }
...
    Error:
      description: Erreur serveur
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        stack:
          type: string
          description: Présent uniquement en environnement development
...
    ClientRequest:
      type: object
      description: Demande stockée sous `Clients/{uid}/Requests/{requestId}` (structure legacy).
      required: [requestId]
      properties:
        requestId:
          type: string
        "User ID":
          type: string
        "Request Type":
          type: string
        "Request Category":
          type: string
        "SubCategory ID":
          type: string
        "Request Sub-Category":
          type: string
        Description:
          type: string
        "Request Date":
          description: Date Firestore (Timestamp) sérialisée
          nullable: true
        Priority:
          type: string
        "Uploaded Files":
          type: array
          items:
            type: string
        "Available Days":
          type: array
          items:
            type: string
        "Available Hours":
          type: array
          items:
            type: string
        Tags:
          type: array
          items:
            type: string
        Status:
          type: string
        "Assigned to":
          type: string
        rating:
          type: number
      additionalProperties: true
    CreateRequestInput:
      type: object
      required: [requestType, category, description]
      properties:
        requestType:
          type: string
        category:
          type: string
        subCategoryId:
          type: string
        subCategory:
          type: string
        description:
          type: string
        priority:
          type: string
          default: normal
        files:
          type: array
          items:
            type: string
        availableDays:
          type: array
          items:
            type: string
        availableHours:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        formData:
          type: object
          additionalProperties: true
...